refactoring: rename is_valid() to check_valid()
A combination of raising and returning a boolean doesn't really make
sense. In the line below the call, one can always assume that the return
value was true (otherwise an exception would have been raised), so it
makes more sense to not return anything and call the function
check_valid().