fix and refactor Profle.add_preferences()
This includes a larger refactoring involving an interface change in the
class Profile. The method add_preferences() is split into two methods,
one to add approvals by one candidate, the other to add many approval
sets at once. This is similar to list.append() and list.extend().

The new implementation separates different semantics in input, to avoid
ambiguous cases (what does [] mean, is it an empty list of approval sets
or an empty list of approvals?). On the other side ducktyping is
supported. The methods accept any iterable of approval sets as input and
each approval set can be a DichotomousPreferences() object or an
iterable of integers. Double votes cause an error.

Note that dictionaries are allowed (they are a valid iterable), one test
case had to be removed.