preparation for refactoring: use objective value
It shouldn't matter if we use the value of the objective function or the
value of the variable we are optimizing for. To be sure, there is an
additional assert.

This will be removed again in the next commit and is just a step to make
code more similar to the other functions in this module.