Convert from verbose flag to Output class
Removes print() calls and replaces them with output.XY() calls. The
verbose flag is not yet removed.

Note that some of the test outputs changed. Mostly due to two reasons:
the verbose flag was not passed on in some places, so verbosity was not
correct before. And propositionA2.py contained calls with different log
level - now everything is set to DEBUG.

This commit is probably quite unmergeable with other changes, let's
simple redo it if there are troubles.