Add unit tests for example scripts
To make things easier the example unit tests were merged with the survey
unit tests. Otherwise we would need to move the fixture to a conftest.py
file and everything gets more complicated.

Note that this commit can be cherry-picked directly on top of a609165326
and tests will succeed with a minor change in the output of simple.py,
so we can assume that we did not break anything major in the examples.